__init__ cleanup and file rename
