+++ svg_turtle/__init__.py
@@ -4,4 +4,4 @@ from .about import __title__, __version__, __url__
 # Importing TurtleGraphicsError directly from turtle will fail without tkinter.
 # noinspection PyUnresolvedReferences
 from .svg_turtle import SvgTurtle, TurtleGraphicsError
-from .IPythonTurtle import IPythonTurtle
+from .ipython_turtle import IPythonTurtle
